Separated grid from main app created second grid based on object array @
